requirements.txt
# Core trading bot dependencies
python-dotenv>=1.0.0
openai>=1.0.0
pyupbit>=0.2.31
pyjwt>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.4.71b0
schedule>=1.2.0
streamlit>=1.28.0
plotly>=5.0.0
requests>=2.28.0

# Chart generation
mplfinance>=0.12.10b0
matplotlib>=3.5.0

# Testing
pytest>=8.0.0
pytest-mock>=3.14.0
